Make sure we only uninstall dphys-swapfile if the platform is Umbrel OS

+++ scripts/update/02-run.sh
@@ -12,8 +12,13 @@ echo "========= Stage: Post-update =========="
 echo "======================================="
 echo
 
-# Remove dphys-swapfile since we now use our own swapfile logic
-if command -v dphys-swapfile >/dev/null 2>&1; then
-  echo "Removing unused dependency \"dphys-swapfile\""
-  apt-get remove -y dphys-swapfile
+# Make Umbrel OS specific post-update changes
+if [[ ! -z "${UMBREL_OS:-}" ]]; then
+
+  # Remove dphys-swapfile since we now use our own swapfile logic
+  if command -v dphys-swapfile >/dev/null 2>&1; then
+    echo "Removing unused dependency \"dphys-swapfile\""
+    apt-get remove -y dphys-swapfile
+  fi
+
 fi
